chore: rename file to match kebab
