+++ packages/monaco/package.json
@@ -9,7 +9,7 @@
     "./vim-mode": "./src/vim-mode.tsx",
     "./twoslash": "./src/twoslash.ts",
     "./prettier": "./src/prettier.ts",
-    "./EditorHooks": "./src/EditorHooks.ts"
+    "./editor-hooks": "./src/editor-hooks.ts"
   },
   "typesVersions": {
     "*": {
@@ -31,8 +31,8 @@
       "prettier": [
         "./src/prettier.ts"
       ],
-      "EditorHooks": [
-        "./src/EditorHooks.ts"
+      "editor-hooks": [
+        "./src/editor-hooks.ts"
       ]
     }
   },
