+++ packages/monaco/src/editor-hooks.ts
